# integrations/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Echo — AI Context Portability API",
    "version": "0.1.0",
    "description": "REST API for Echo's universal AI context portability layer. Save, load, and manage portable context that follows the user across AI tools — encrypted, stored on Filecoin, controlled by the user's smart contract."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/renewal/fund": {
      "post": {
        "operationId": "fundRenewal",
        "summary": "Fund perpetual-storage renewal endowment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amountInFil"],
                "properties": {
                  "amountInFil": {
                    "type": "string",
                    "description": "Amount of FIL to fund (e.g. '0.1')"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Renewal funded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "funded": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}